+++ src/main/java/web/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> getUsers();
-    public User showUser(int id);
+    public User showUser(Long id);
     public void saveUser(User user);
-    public void updateUser(int id,User user);
-    public void deleteUser(int id);
+    public void updateUser(Long id,User user);
+    public void deleteUser(Long id);
 }
